chore(pom.xml): update maven-failsafe-plugin configuration to include integration tests
The maven-failsafe-plugin configuration in the pom.xml file has been updated to include the execution of integration tests. The skipITs property has been replaced with includes property, which specifies the patterns for the integration test files to be executed. This change allows the integration tests with filenames starting with "IT" or ending with "IT" to be included in the test execution.